Extracted 'name by item' and 'item by name' from structs.h
