+++ src/engine/entities/entities_constants.h
@@ -9,6 +9,7 @@
 #define BYLINS_SRC_ENTITY_ROOMS_ROOM_CONSTANTS_H_
 
 #include "engine/structs/structs.h"
+#include "engine/structs/meta_enum.h"
 
 /*
  * ========================================================================================
