--- src/engine/structs/structs.cpp
@@ -1,6 +1,0 @@
-
-const char* nothing_string = "ничего";
-
-
-// vim: ts=4 sw=4 tw=0 noet syntax=cpp :
-
